requirements.txt
torch
transformers
datasets
peft
accelerate
bitsandbytes
evaluate
scikit-learn
google-generativeai